Stochastic C++ model output values (with treatment applied) are now similar to those of the Deterministic R model.

+++ CRealization.h
@@ -67,7 +67,7 @@ public:
 	double freeliving;
 	double* compareArray;
 	double* hostTotalWorms;
-	double* productiveFemaleWorms;
+	int* productiveFemaleWorms;
 	double* eggsOutputPerHost;
 	double eggsProductionRate;
 	double* hostInfectionRate;
